fix module versioning with our current approach
