+++ backend/src/services/identity-v2/identity-service.ts
@@ -1,20 +1,33 @@
-import { AccessScope, OrgMembershipRole } from "@app/db/schemas";
+import { AccessScope, ActionProjectType, OrganizationActionScope, OrgMembershipRole } from "@app/db/schemas";
 import { TLicenseServiceFactory } from "@app/ee/services/license/license-service";
+import {
+  OrgPermissionAdminConsoleAction,
+  OrgPermissionIdentityActions,
+  OrgPermissionSubjects
+} from "@app/ee/services/permission/org-permission";
 import { TPermissionServiceFactory } from "@app/ee/services/permission/permission-service-types";
+import { ProjectPermissionIdentityActions, ProjectPermissionSub } from "@app/ee/services/permission/project-permission";
 import { TKeyStoreFactory } from "@app/keystore/keystore";
 import { BadRequestError, NotFoundError } from "@app/lib/errors";
-import { getIdentityActiveLockoutAuthMethods } from "@app/services/identity-v2/identity-fns";
+import { TProjectDALFactory } from "@app/services/project/project-dal";
 
+import { ActorType } from "../auth/auth-type";
+import { getIdentityActiveLockoutAuthMethods } from "../identity/identity-fns";
 import { TIdentityMetadataDALFactory } from "../identity/identity-metadata-dal";
 import { TIdentityAccessTokenServiceFactory } from "../identity-access-token/identity-access-token-service";
 import { TMembershipRoleDALFactory } from "../membership/membership-role-dal";
 import { TMembershipIdentityDALFactory } from "../membership-identity/membership-identity-dal";
 import { TIdentityV2DALFactory } from "./identity-dal";
+import { filterIdentitiesByProjectPermission, TProjectPermissionAbility } from "./identity-fns";
+import { TIdentityMembershipV2DALFactory } from "./identity-membership-dal";
 import {
+  SearchIdentitiesScope,
+  TCountIdentitiesV2DTO,
   TCreateIdentityV2DTO,
   TDeleteIdentityV2DTO,
   TGetIdentityByIdV2DTO,
   TListIdentityV2DTO,
+  TSearchIdentitiesV2DTO,
   TUpdateIdentityV2DTO
 } from "./identity-types";
 import { newOrgIdentityFactory } from "./org/org-identity-factory";
@@ -22,6 +35,7 @@ import { newProjectIdentityFactory } from "./project/project-identity-factory";
 
 type TScopedIdentityV2ServiceFactoryDep = {
   identityDAL: TIdentityV2DALFactory;
+  identityMembershipV2DAL: TIdentityMembershipV2DALFactory;
   permissionService: TPermissionServiceFactory;
   licenseService: Pick<TLicenseServiceFactory, "getPlan" | "updateSubscriptionOrgMemberCount">;
   membershipIdentityDAL: TMembershipIdentityDALFactory;
@@ -29,19 +43,22 @@ type TScopedIdentityV2ServiceFactoryDep = {
   identityMetadataDAL: TIdentityMetadataDALFactory;
   identityAccessTokenService: Pick<TIdentityAccessTokenServiceFactory, "revokeAllTokensForIdentity">;
   keyStore: Pick<TKeyStoreFactory, "getKeysByPattern" | "getItem">;
+  projectDAL: Pick<TProjectDALFactory, "findActorAccessibleProjectIds" | "findOrgProjectIds">;
 };
 
 export type TScopedIdentityV2ServiceFactory = ReturnType<typeof identityV2ServiceFactory>;
 
 export const identityV2ServiceFactory = ({
   identityDAL,
+  identityMembershipV2DAL,
   permissionService,
   licenseService,
   membershipIdentityDAL,
   membershipRoleDAL,
   identityMetadataDAL,
   identityAccessTokenService,
-  keyStore
+  keyStore,
+  projectDAL
 }: TScopedIdentityV2ServiceFactoryDep) => {
   const orgFactory = newOrgIdentityFactory({
     permissionService
@@ -249,11 +266,242 @@ export const identityV2ServiceFactory = ({
     return { ...identities, docs: identities.docs.filter((el) => isIdentityAccessible({ identityId: el.id })) };
   };
 
+  const resolveIdentitySearchScope = async ({
+    actor,
+    actorId,
+    actorAuthMethod,
+    actorOrgId,
+    scope
+  }: {
+    actor: ActorType;
+    actorId: string;
+    actorAuthMethod: TSearchIdentitiesV2DTO["actorAuthMethod"];
+    actorOrgId: string;
+    scope: SearchIdentitiesScope[];
+  }): Promise<{
+    uniqueScope: Set<SearchIdentitiesScope>;
+    accessibleProjectIds: string[];
+    projectPermissions: Map<string, TProjectPermissionAbility>;
+    conditionalProjectIds: Set<string>;
+  }> => {
+    const uniqueScope = new Set(scope);
+    const { permission: orgPermission } = await permissionService.getOrgPermission({
+      scope: OrganizationActionScope.Any,
+      actor,
+      actorId,
+      orgId: actorOrgId,
+      actorAuthMethod,
+      actorOrgId
+    });
+    const canReadOrgIdentities = orgPermission.can(OrgPermissionIdentityActions.Read, OrgPermissionSubjects.Identity);
+    if (uniqueScope.has(SearchIdentitiesScope.OrganizationScope) && !canReadOrgIdentities) {
+      uniqueScope.delete(SearchIdentitiesScope.OrganizationScope);
+    }
+
+    const accessibleProjectIds: string[] = [];
+    const projectPermissions = new Map<string, TProjectPermissionAbility>();
+    const conditionalProjectIds = new Set<string>();
+    if (uniqueScope.has(SearchIdentitiesScope.ProjectScope)) {
+      const canAccessAllProjects = orgPermission.can(
+        OrgPermissionAdminConsoleAction.AccessAllProjects,
+        OrgPermissionSubjects.AdminConsole
+      );
+
+      if (canAccessAllProjects && canReadOrgIdentities) {
+        // Org admins read unconditionally across every project — skip the per-project probe,
+        // which would otherwise throw ProjectMembershipNotFound for projects the admin hasn't
+        // explicitly joined and drop them from accessibleProjectIds. The org-level identity:read
+        // gate stops access-all-projects holders without identity:read from enumerating
+        // project-scoped machine identities across the org.
+        accessibleProjectIds.push(...(await projectDAL.findOrgProjectIds(actorOrgId)));
+      } else {
+        const candidateProjectIds = await projectDAL.findActorAccessibleProjectIds(actorId, actor, actorOrgId);
+
+        const projectAccessChecks = await Promise.all(
+          candidateProjectIds.map(async (projectId) => {
+            try {
+              const { permission: projectPermission } = await permissionService.getProjectPermission({
+                actor,
+                actorId,
+                actionProjectType: ActionProjectType.Any,
+                actorAuthMethod,
+                projectId,
+                actorOrgId
+              });
+              // Broad `can(Read, Identity)` returns true for conditional rules too. We keep the
+              // project here and use the per-row check below to honor `identityId` conditions.
+              return projectPermission.can(ProjectPermissionIdentityActions.Read, ProjectPermissionSub.Identity)
+                ? { projectId, permission: projectPermission }
+                : null;
+            } catch {
+              return null;
+            }
+          })
+        );
+
+        for (const entry of projectAccessChecks) {
+          if (entry) {
+            accessibleProjectIds.push(entry.projectId);
+            projectPermissions.set(entry.projectId, entry.permission);
+            // Tracking which projects carry conditional Read(Identity) rules lets the row filter
+            // skip CASL.can() for projects whose access is already unconditional — the broad
+            // can(Read, Identity) check above is authoritative for them.
+            const rules = entry.permission.rulesFor(
+              ProjectPermissionIdentityActions.Read,
+              ProjectPermissionSub.Identity
+            );
+            if (rules.some((rule) => rule.conditions)) {
+              conditionalProjectIds.add(entry.projectId);
+            }
+          }
+        }
+      }
+
+      if (accessibleProjectIds.length === 0) {
+        uniqueScope.delete(SearchIdentitiesScope.ProjectScope);
+      }
+    }
+
+    return { uniqueScope, accessibleProjectIds, projectPermissions, conditionalProjectIds };
+  };
+
+  const searchOrgIdentities = async ({
+    actor,
+    actorId,
+    actorAuthMethod,
+    actorOrgId,
+    limit,
+    offset,
+    orderBy,
+    orderDirection,
+    scope,
+    searchFilter = {}
+  }: TSearchIdentitiesV2DTO) => {
+    const { uniqueScope, accessibleProjectIds, projectPermissions, conditionalProjectIds } =
+      await resolveIdentitySearchScope({
+        actor,
+        actorId,
+        actorAuthMethod,
+        actorOrgId,
+        scope
+      });
+
+    if (!uniqueScope.size) {
+      return { identityMemberships: [], totalCount: 0 };
+    }
+
+    if (conditionalProjectIds.size === 0) {
+      const { totalCount, docs } = await identityMembershipV2DAL.searchIdentitiesV2({
+        orgId: actorOrgId,
+        limit,
+        offset,
+        orderBy,
+        orderDirection,
+        searchFilter,
+        scope: Array.from(uniqueScope),
+        accessibleProjectIds
+      });
+
+      return { identityMemberships: docs, totalCount };
+    }
+
+    // Conditional Read rules on Identity exist for at least one accessible project. SQL pagination
+    // would over-return rows the caller can't see, so fetch the full ordered set and filter by
+    // per-row CASL before paginating in memory. Mirrors the tier-2 pattern in pam-resource-service.
+    const { docs } = await identityMembershipV2DAL.searchIdentitiesV2({
+      orgId: actorOrgId,
+      orderBy,
+      orderDirection,
+      searchFilter,
+      scope: Array.from(uniqueScope),
+      accessibleProjectIds
+    });
+
+    const filtered = filterIdentitiesByProjectPermission(docs, projectPermissions, conditionalProjectIds);
+    const pageOffset = offset ?? 0;
+    const pageLimit = limit ?? filtered.length;
+    return {
+      identityMemberships: filtered.slice(pageOffset, pageOffset + pageLimit),
+      totalCount: filtered.length
+    };
+  };
+
+  const countOrgIdentities = async ({
+    actor,
+    actorId,
+    actorAuthMethod,
+    actorOrgId,
+    scope,
+    searchFilter = {}
+  }: TCountIdentitiesV2DTO) => {
+    const requestedOrg = scope.includes(SearchIdentitiesScope.OrganizationScope);
+    const requestedProject = scope.includes(SearchIdentitiesScope.ProjectScope);
+
+    const { uniqueScope, accessibleProjectIds, projectPermissions, conditionalProjectIds } =
+      await resolveIdentitySearchScope({
+        actor,
+        actorId,
+        actorAuthMethod,
+        actorOrgId,
+        scope
+      });
+
+    // Build the response shape from the *requested* scopes so callers always get a number
+    // for every scope they asked about (zero when permissions/projects filter the scope out).
+    const counts: { organization?: number; project?: number } = {};
+    if (requestedOrg) counts.organization = 0;
+    if (requestedProject) counts.project = 0;
+
+    if (!uniqueScope.size) {
+      return counts;
+    }
+
+    if (conditionalProjectIds.size === 0) {
+      const dalCounts = await identityMembershipV2DAL.countIdentitiesV2({
+        orgId: actorOrgId,
+        scope: Array.from(uniqueScope),
+        accessibleProjectIds,
+        searchFilter
+      });
+
+      if (requestedOrg && dalCounts.organization !== undefined) counts.organization = dalCounts.organization;
+      if (requestedProject && dalCounts.project !== undefined) counts.project = dalCounts.project;
+
+      return counts;
+    }
+
+    // Project-scope count must respect identityId conditions, so list the rows and count survivors.
+    // Org-scope rows have no per-identity conditions, so the SQL count is still correct.
+    if (requestedOrg && uniqueScope.has(SearchIdentitiesScope.OrganizationScope)) {
+      const orgOnly = await identityMembershipV2DAL.countIdentitiesV2({
+        orgId: actorOrgId,
+        scope: [SearchIdentitiesScope.OrganizationScope],
+        accessibleProjectIds: [],
+        searchFilter
+      });
+      counts.organization = orgOnly.organization ?? 0;
+    }
+
+    if (requestedProject && uniqueScope.has(SearchIdentitiesScope.ProjectScope)) {
+      const { docs } = await identityMembershipV2DAL.searchIdentitiesV2({
+        orgId: actorOrgId,
+        scope: [SearchIdentitiesScope.ProjectScope],
+        accessibleProjectIds,
+        searchFilter
+      });
+      counts.project = filterIdentitiesByProjectPermission(docs, projectPermissions, conditionalProjectIds).length;
+    }
+
+    return counts;
+  };
+
   return {
     createIdentity,
     updateIdentity,
     deleteIdentity,
     getIdentityById,
-    listIdentities
+    listIdentities,
+    searchOrgIdentities,
+    countOrgIdentities
   };
 };
